refactor: rename node-patches to node-bootstrap, register.cjs to bootstrap.cjs
This is a pure rename in preparation for a followup change that will move more
launcher functionality into this file, since it will do more than patch fs
going forward. Eventually we would like to get rid of the bash launcher
altogether.
