+++ .prettierignore
@@ -4,7 +4,7 @@ examples/**/*-docs.md
 examples/**/*_pb.d.ts
 js/private/coverage/coverage.js
 js/private/devserver/js_run_devserver.mjs
-js/private/node-patches/fs.cjs
+js/private/node-bootstrap/fs.cjs
 js/private/watch/aspect_watch_protocol.mjs
 js/private/watch/aspect_watch_protocol.d.mts
 min/
