+++ js/private/js_binary.bzl
@@ -263,11 +263,11 @@ _ATTRS = {
     "_windows_constraint": attr.label(default = "@platforms//os:windows"),
     "_node_patches_files": attr.label_list(
         allow_files = True,
-        default = [Label("@aspect_rules_js//js/private/node-patches:fs.cjs")],
+        default = [Label("@aspect_rules_js//js/private/node-bootstrap:fs.cjs")],
     ),
     "_node_patches": attr.label(
         allow_single_file = True,
-        default = Label("@aspect_rules_js//js/private/node-patches:register.cjs"),
+        default = Label("@aspect_rules_js//js/private/node-bootstrap:bootstrap.cjs"),
     ),
 }
