+++ js/private/node-bootstrap/fs.cjs
@@ -1,4 +1,4 @@
-// Generated by //js/private/node-patches:compile
+// Generated by //js/private/node-bootstrap:compile
 "use strict";
 /**
  * @license
