+++ js/private/node-bootstrap/src/BUILD.bazel
@@ -22,6 +22,6 @@ genrule(
     name = "fs_generated",
     srcs = ["fs.cjs"],
     outs = ["fs-generated.cjs"],
-    cmd = "echo \"// Generated by //js/private/node-patches:compile\" > $@ && cat $(execpath :fs.cjs) >> $@",
-    visibility = ["//js/private/node-patches:__pkg__"],
+    cmd = "echo \"// Generated by //js/private/node-bootstrap:compile\" > $@ && cat $(execpath :fs.cjs) >> $@",
+    visibility = ["//js/private/node-bootstrap:__pkg__"],
 )
