+++ js/private/node-bootstrap/src/fs.cts
