+++ js/private/test/image/checksum_test.expected
@@ -1,4 +1,4 @@
-9d11f575b6217b8f310772a693263d24fd56e6c0d4d98792f4fc84e8ceeec307  js/private/test/image/cksum_node.tar
+031bbb0850b697e75e647ab254e4de6511ffd7a249b4b48ca719cb4d8d9c4783  js/private/test/image/cksum_node.tar
 70b10220a2c05d87da4271c17c38ded8febc725e20e06f1c3809d2bb02ba4ae7  js/private/test/image/cksum_package_store_3p.tar
 2cb6f678d6eb0b2e9d5e2637f41ae3f192233752f1ea2a55cede2531deec2a64  js/private/test/image/cksum_package_store_1p.tar
 79afa99006aff19460354e72cf2634db693670d09ccc7531fbf27f1f20fe0a5f  js/private/test/image/cksum_node_modules.tar
